Make examples/VC709 work
The VC709 and NetFPGA_SUME are very similar.  The same gtx transceiver
IP blocks are used, just with different pin assigmnents.

The I2C MUX and SI5324 resets are active low on the VC709 compared
to active high on the NetFPGA_SUME.  Additionally, the SI5324
is connected to port 7 of the I2C mux on VC709 compared to being
connected to port 4 on the NetFPGA.

Other than that the LEDs and buttons are different.