Add Xilinx Kintex UltraScale+ KCU116 board
This commit adds the Xilinx Kintex UltraScale+ KCU116 board featuring
the XCKU5P-2FFVB676E FPGA and 4 zSFP cages. XGMII PHYs are
instantiated for all four cages in `fpga.v` and passed down to
`fpga_core.v`. Only the first (SFP 0) cage is connected to the
instantiated UDP/IP core.

The I2C bus to the SFPs uses an I2C mux chip and is currently
unsupported.

The reference clock for the GTY transceiver connected to the SFPs is
generated by an external, on-board Si5328 clock chip which must be
configured to output a clock signal of 156.25MHz before the
transceiver can be used. This configuration is performed through the
onboard Zynq-based baseboard management controller. Software for
configuring the Zynq BMC and setting the proper clock frequencies can
be found on Xilinx' website.